pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bubblemask"
version = "1.0.0"
authors = [
  { name="Jack E. Taylor", email="Taylor@psych.uni-frankfurt.de" },
]
description = "Package for applying Gaussian 'Bubbles' masks to images."
readme = "README.md"
dependencies = [
  "numpy>=2.0",
  "Pillow>=10.0.1",
  "scikit-image>=0.19.3",
  "scipy>=1.13.0"
]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
]

[project.urls]
Homepage = "https://github.com/JackEdTaylor/bubblemask"
Issues = "https://github.com/JackEdTaylor/bubblemask/issues"
